requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
supabase==2.0.2
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI & ML
openai==1.3.7
anthropic==0.7.8

# HTTP & Networking
httpx==0.25.2
aiohttp==3.9.1

# Validation & Serialization
pydantic[email]==2.5.0
email-validator==2.1.0

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0

# Environment & Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
prometheus-client==0.19.0

